Preprocess: handle new Ranges placeholder links
The site uses placeholder links that get replaced by JS to point to
either Ranges TS or standard Ranges pages, depending on whether they are
used on Ranges TS pages or not. This is necessary because the links are
generated by the syntax highlighter, which doesn't know about the
different targets. This change adds handling of those placeholder links
to replace them by proper relative links.