Preprocess: improve cssless processing speed
This greatly improves the speed of cssless processing by using regular
expressions instead of a full-blown CSS parser to manipulate the `style`
attribute where possible. The CSS used on the site is simple enough so
this works just fine.

The tests needed some massaging to get them to pass again, because the
new handling caused the following minor changes to the output.

 * Floating point values are always formatted with the decimal point,
   which was omitted previously when the fractional part was zero.
 * Existing CSS declarations are not touched when other declarations are
   added or removed, so there are some differences where spaces were
   added previously but aren't now.