Preprocess: add premailer source code
Premailer is used by the cssless preprocessing script to inline CSS
styles, it hasn't been maintained in a while. This adds the source code
of the premailer module so we can make performance improvements to
cssless preprocessing. These are going to be changes specific to this
project, which won't be of any use to upstream premailer.