Preprocess: further improve cssless speed
This further improves the speed of cssless processing by using regular
expressions in premailer instead of full parsing, if possible. Premailer
is now also used with validation turned off, we just assume that the CSS
we get from the site is valid.

This takes the time for running unit tests from over 7 seconds before
the improvements down to about 1.4 seconds, which should make building a
release more bearable (keep in mind we're talking about 5k files to be
processed, each taking about 0.7 seconds).