Preprocess: simplify file renaming
This simplifies file renaming by using a map from old to new name. The
new rename map doesn't store file paths any more, so the directory tree
needs to be walked twice (once for searching, once for renaming), but
transforming relative links is a lot simpler because it can use a map
lookup instead of trying every entry in the rename map.
Transformation of relative links now uses urllib for parsing URLs, so
matching and transformation should be more robust for exotic links.